# docs/swagger.yaml
openapi: 3.0.0
servers:
  - url: localhost:25565
info:
  description: An API used to do banking operations such as user register, withdrawl, deposit, access control, internal and external accounts transfers, account blocking and more.
  version: "1.0.0"
  title: "Simple Banking API"
  contact:
    email: julio.kobayashi@ebury.com
  license:
    name: Apache 2.0
    url: "[http://www.apache.org/licenses/LICENSE-2.0.html](http://www.apache.org/licenses/LICENSE-2.0.html)"
...
paths:
...
  /accounts:
    post:
      tags:
        - accounts
      summary: creates an account
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Account'
      responses:
        "201":
          description: Account created sucessfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Account'
        "400":
          description: bad input parameter
        "403":
          description: Forbidden
        "500":
          description: Internal Server Error
  /accounts/{account_id}:
    get:
      tags:
        - accounts
      summary: return data of account with ID accountID
      parameters:
        - in: path
          name: account_id
          required: true
          schema:
            type: string
      responses:
        "200":
          description: account returned sucessfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Account'
        "401":
          description: Unauthorized
        "403":
          description: Forbidden
        "500":
          description: Internal Server Error
    delete:
      tags:
        - accounts
      summary: delete account of accountID
      parameters:
        - in: path
          name: account_id
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Account Deleted
        "401":
          description: Unauthorized
        "403":
          description: Forbidden
        "500":
          description: Internal Server Error
    put:
      tags:
        - accounts
      summary: update a given piece of information of a certain account
      parameters:
        - in: path
          name: account_id
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Account'
      responses:
        "200":
          description: Account updated sucessfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Account'
        "401":
          description: Unauthorized
        "403":
          description: Forbidden
        "500":
          description: Internal Server Error
  /accounts/report/{account_id}:
    get:
      tags:
        - accounts
      summary: returns a report of the account
      parameters:
        - in: path
          name: account_id
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Report returned
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AccountReport'
        "401":
          description: Unauthorized
        "403":
          description: Forbidden
        "500":
          description: Internal Server Error
  /accounts/{account_id}/block:
    put:
      tags:
        - accounts
      summary: block an account
      parameters:
        - in: path
          name: account_id
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Account Blocked Sucesfully!
        "401":
          description: Unauthorized
        "403":
          description: Forbidden
        "500":
          description: Internal Server Error
  /accounts/{account_id}/unblock:
    put:
      tags:
        - accounts
      summary: unblock an account
      parameters:
        - in: path
          name: account_id
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Account Unblocked
        "401":
          description: Unauthorized
        "403":
          description: Forbidden
        "500":
          description: Internal Server Error
...
  /autodebits:
    post:
      tags:
        - automaticdebits
      summary: creates a automatic debit
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AutomaticDebit'
      responses:
        "202":
          description: Automatic debit created sucessfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AutomaticDebit'
        "401":
          description: Unauthorized
        "403":
          description: Forbidden
        "500":
          description: Internal Server Error
  /autodebits/{debit_id}:
    get:
      tags:
        - automaticdebits
      summary: return data of automatic debit with ID debit_id
      parameters:
        - in: path
          name: debit_id
          required: true
          schema:
            type: string
      responses:
        "200":
          description: automatic debit returned sucessfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AutomaticDebit'
        "401":
          description: Unauthorized
        "403":
          description: Forbidden
        "500":
          description: Internal Server Error
    delete:
      tags:
        - automaticdebits
      summary: delete automatic debit of debit_id
      parameters:
        - in: path
          name: debit_id
          required: true
          schema:
            type: string
      responses:
        "204":
          description: Automatic debit deleted successfully
        "401":
          description: Unauthorized
        "403":
          description: Forbidden
        "500":
          description: Internal Server Error
components:
  schemas:
...
    Account:
      type: object
      properties:
        account_id:
          type: string
        client_id:
          type: string
        user_id:
          type: string
        agency_id:
          type: integer
          format: uint32
        register_date:
          type: string
          format: date-time
        balance:
          type: number
          format: float64
        status:
          type: boolean
    Transfer:
      type: object
      properties:
        transfer_id:
          type: string
        account_id:
          type: string
        account_to:
          type: string
        value:
          type: number
          format: float64
        register_date:
          type: string
          format: date-time
    Deposit:
      type: object
      properties:
        deposit_id:
          type: string
        account_id:
          type: string
        client_id:
          type: string
        user_id:
          type: string
        agency_id:
          type: integer
          format: uint32
        deposit:
          type: number
          format: float64
        deposit_date:
          type: string
          format: date-time
    Withdrawal:
      type: object
      properties:
        withdrawal_id:
          type: string
        account_id:
          type: string
        client_id:
          type: string
        agency_id:
          type: integer
          format: uint32
        withdrawal:
          type: number
          format: float64
        withdrawal_date:
          type: string
          format: date-time
        status:
          type: boolean
    AutomaticDebit:
      type: object
      properties:
        debit_id:
          type: string
        account_id:
          type: string
        client_id:
          type: string
        agency_id:
          type: integer
          format: uint32
        value:
          type: number
          format: float64
        debit_day:
          type: integer
          format: uint16
        status:
          type: boolean
        expiration_date:
          type: string
          format: date-time
        register_date:
          type: string
          format: date-time
...
    AccountReport:
      type: object
      properties:
        account_id:
          type: string
        client_id:
          type: string
        agency_id:
          type: integer
          format: uint32
        balance:
          type: number
          format: float64
        register_date:
          type: string
          format: date-time
        status:
          type: boolean
        transfers:
          type: array
          items:
            $ref: '#/components/schemas/Transfer'
        deposits:
          type: array
          items:
            $ref: '#/components/schemas/Deposit'
        withdrawals:
          type: array
          items:
            $ref: '#/components/schemas/Withdrawal'
        automatic_debits:
          type: array
          items:
            $ref: '#/components/schemas/AutomaticDebit'
        report_date:
          type: string
          format: date-time